backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
reportlab==4.0.7
supabase==2.3.0
python-dotenv==1.0.0
